requirements.txt
# Azure Dependencies
azure-identity>=1.12.0
azure-mgmt-compute>=29.1.0
azure-mgmt-network>=23.1.0
azure-core>=1.26.4

# GCP Dependencies
google-cloud-compute>=2.0.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0
google-api-python-client>=2.0.0

# Common Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
logging>=0.4.9
